feat: add phase system and version bump
